# apps/api/openapi/openapi.yaml
openapi: 3.0.2
info:
  version: 1.0.0
  title: CaenCamp.s JobBoard API
  description: "Description de l'API Rest des offres d'emplois des CaenCamp.s"
  termsOfService: 'https://jobboard.caen.camp/terms/'
  contact:
    name: JobBoard Support
    url: 'https://jobboard.caen.camp/support'
    email: jobboard@caen.camp
  license:
    name: GNU GPLv3
    url: 'https://github.com/CaenCamp/jobs-caen-camp/blob/master/LICENSE'
servers:
  - url: 'http://localhost:8001'
    description: Local server
security: []
...
paths:
...
  /api/organizations:
    get:
      summary: Liste des entreprises
      tags:
        - Entreprises
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FullOrganization'
              examples:
                example-1:
                  value:
                    - id: a122edec-5580-4a93-aff7-fc18b41e4c57
                      name: Incaya
                      description: "INCAYA, société coopérative de services du numérique. Créons ensemble des outils qui ont du sens ! Collectivités, associations, entreprises : nous concevons avec vous des outils numériques qui transforment l'expérience de vos utilisateurs et usagers."
                      image: 'https://www.incaya.fr/static/logo-incaya.svg'
                      email: contact@organization.org
                      url: 'https://www.incaya.fr/'
                      address:
                        addressCountry: France
                        addressLocality: Caen
                        postalCode: '14000'
                        streetAddress: '5, place de la Répulique'
                      contactPoints:
                        - email: job@organization.org
                          telephone: '0606060606'
                          name: 'John Do, CTO'
                          contactType: "Offres d'emploi"
          headers:
            X-Total-Count:
                $ref: '#/components/headers/X-Total-Count'
            Link:
                $ref: '#/components/headers/Link'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExtendedErrorModel'
      operationId: getOrganizations
      description: "Obtenir la liste des entreprises.\n\nCette liste est filtrable par filters={ key: value, key2: value ...} :\n* name\n* address_locality\n* postal_code\n\nCette liste est triable par sort=[key, sortDirection_ASC_DESC] : \n* name\n* address_locality\n* postal_code\n\nLa pagination est controlable via le paramètre de requête pagination=[nombreDeRésultatsParPage, numeroDePage]"
      parameters:
        - $ref: '#/components/parameters/Filter'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/PaginationCurrentPage'
        - $ref: '#/components/parameters/PaginationPerPage'
    post:
      summary: "Création d'une entreprise"
      operationId: createOrganization
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FullOrganization'
              examples:
                example-1:
                  value:
                    id: a122edec-5580-4a93-aff7-fc18b41e4c57
                    name: Incaya
                    description: "INCAYA, société coopérative de services du numérique. Créons ensemble des outils qui ont du sens ! Collectivités, associations, entreprises : nous concevons avec vous des outils numériques qui transforment l'expérience de vos utilisateurs et usagers."
                    image: 'https://www.incaya.fr/static/logo-incaya.svg'
                    email: contact@organization.org
                    url: 'https://www.incaya.fr/'
                    address:
                      addressCountry: France
                      addressLocality: Caen
                      postalCode: '14000'
                      streetAddress: '5, place de la Répulique'
                    contactPoints:
                      - email: job@organization.org
                        telephone: '0606060606'
                        name: 'John Do, CTO'
                        contactType: "Offres d'emploi"
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExtendedErrorModel'
      description: Créer une entreprise
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationForCreation'
            examples:
              example-1:
                value:
                  description: desc
                  image: 'https://www.org.org/logo.svg'
                  email: contact@org.org
                  url: 'https://www.org.org'
                  address:
                    addressCountry: FR
                    addressLocality: Caen
                    postalCode: '14000'
                    streetAddress: '5, place de la Répulique'
                  contactPoints:
                    - email: job@org.org
                      telephone: '0606060606'
                      name: 'John Do, CTO'
                      contactType: 'Offres d"emploi'
      tags:
        - Entreprises
    parameters: []
  '/api/organizations/{identifier}':
    parameters:
      - $ref: '#/components/parameters/UUID'
    get:
      summary: "Donnée d'une entreprise"
      tags:
        - Entreprises
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FullOrganization'
              examples:
                example-1:
                  value:
                    id: a122edec-5580-4a93-aff7-fc18b41e4c57
                    name: Incaya
                    description: "INCAYA, société coopérative de services du numérique. Créons ensemble des outils qui ont du sens ! Collectivités, associations, entreprises : nous concevons avec vous des outils numériques qui transforment l'expérience de vos utilisateurs et usagers."
                    image: 'https://www.incaya.fr/static/logo-incaya.svg'
                    email: contact@organization.org
                    url: 'https://www.incaya.fr/'
                    address:
                      addressCountry: France
                      addressLocality: Caen
                      postalCode: '14000'
                      streetAddress: '5, place de la Répulique'
                    contactPoints:
                      - email: job@organization.org
                        telephone: '0606060606'
                        name: 'John Do, CTO'
                        contactType: "Offres d'emploi"
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExtendedErrorModel'
      operationId: getOrganization
      description: "Obtenir les données d'une entreprise"
    put:
      summary: Mettre à jour une entreprise
      operationId: updateOrganization
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FullOrganization'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExtendedErrorModel'
      description: "Mettre à jour les données d'une entreprise"
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Organization'
            examples:
              example-1:
                value:
                  id: a122edec-5580-4a93-aff7-fc18b41e4c57
                  name: Incaya
                  description: "INCAYA, société coopérative de services du numérique. Créons ensemble des outils qui ont du sens ! Collectivités, associations, entreprises : nous concevons avec vous des outils numériques qui transforment l'expérience de vos utilisateurs et usagers."
                  image: 'https://www.incaya.fr/static/logo-incaya.svg'
                  email: contact@organization.org
                  url: 'https://www.incaya.fr/'
                  address:
                    addressCountry: France
                    addressLocality: Caen
                    postalCode: '14000'
                    streetAddress: '5, place de la Répulique'
                  contactPoints:
                    - email: job@organization.org
                      telephone: '0606060606'
                      name: 'John Do, CTO'
                      contactType: "Offres d'emploi"
        description: "Les données de l'entreprise mises à jour"
      tags:
        - Entreprises
    delete:
      summary: "Suppression d'une entreprise"
      operationId: deleteOrganization
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                    example: 9a6c8995-df54-446c-a5b8-71532c304751
                required:
                  - id
              examples:
                example-1:
                  value:
                    id: 9a6c8995-df54-446c-a5b8-71532c304751
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExtendedErrorModel'
      description: Supprimer une entreprise existante
      tags:
        - Entreprises
components:
  securitySchemes: {}
  headers:
    X-Total-Count:
        description: "Return information about total items to paginate"
        schema:
            type: integer
        example: 42
        required: false
    Link:
        description: "Return information about pagination actions"
        schema:
            type: string
        example: |
            </resources?currentPage=3&perPage=10>; rel="self",
            </resources?currentPage=1&perPage=10>; rel="first",
            </resources?currentPage=2&perPage=10>; rel="prev",
            </resources?currentPage=4&perPage=10>; rel="next",
            </resources?currentPage=5&perPage=10>; rel="last",
        required: false
  schemas:
...
    Organization:
      description: Une entreprise liée au développement informatique.
      type: object
      x-examples:
        example-hiringOrganization:
          name: Incaya
          description: "INCAYA, société coopérative de services du numérique. Créons ensemble des outils qui ont du sens ! Collectivités, associations, entreprises : nous concevons avec vous des outils numériques qui transforment l'expérience de vos utilisateurs et usagers."
          image: 'https://www.incaya.fr/static/logo-incaya.svg'
          email: contact@organization.org
          url: 'https://www.incaya.fr/'
          address:
            addressCountry: France
            addressLocality: Caen
            postalCode: '14000'
            streetAddress: '5, place de la Répulique'
          contactPoints:
            - identifier: abeb9875-2191-4097-8249-5b8defc42a3e
              email: job@organization.org
              telephone: '0606060606'
              name: 'John Do, CTO'
              contactType: "Offres d'emploi"
      properties:
        name:
          type: string
          example: Incaya
          minLength: 2
          maxLength: 150
        description:
          type: string
          example: Coopérative de services du numérique
        url:
          type: string
          format: uri
          example: 'https://www.incaya.fr/'
          minLength: 10
          pattern: "https?:\\/\\/(www\\.)?[-a-zA-Z0-9@:%._\\+~#=]{1,256}\\.[a-zA-Z0-9()]{1,6}\\b([-a-zA-Z0-9()@:%_\\+.~#?&//=]*)"
        email:
          type: string
          format: email
          example: email@organization.org
        image:
          type: string
          format: uri
          example: 'https://www.incaya.fr/static/logo-incaya.svg'
          minLength: 10
          pattern: "https?:\\/\\/(www\\.)?[-a-zA-Z0-9@:%._\\+~#=]{1,256}\\.[a-zA-Z0-9()]{1,6}\\b([-a-zA-Z0-9()@:%_\\+.~#?&//=]*)"
        address:
          anyOf:
            - type: string
            - $ref: '#/components/schemas/PostalAddress'
        contactPoints:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/FullContactPoint'
      required:
        - name
        - url
    PostalAddress:
      title: PostalAddress
      type: object
      properties:
        addressCountry:
          type: string
          example: France
          minLength: 2
        addressLocality:
          type: string
          example: Caen
          minLength: 2
        postalCode:
          type: string
          example: '14000'
          minLength: 5
          maxLength: 5
        streetAddress:
          type: string
          example: '5, place de la Répulique'
      required:
        - addressLocality
        - postalCode
      description: Une adresse postale.
      x-examples:
        example-Caen:
          addressCountry: France
          addressLocality: Caen
          postalCode: '14000'
          streetAddress: '5, place de la Répulique'
    ContactPoint:
      title: ContactPoint
      type: object
      x-examples:
        example-job:
          email: job@organization.org
          telephone: '0606060606'
          name: 'John Do, CTO'
          contactType: "Offres d'emploi"
      description: "Une personne à contacter au sein de l'organisation"
      properties:
        email:
          type: string
          format: email
          example: job@organization.org
          minLength: 7
        telephone:
          type: string
          nullable: true
        name:
          type: string
        contactType:
          type: string
          example: "Offres d'emploi"
          nullable: true
      required:
        - email
        - name
    ExtendedErrorModel:
      allOf:
        - $ref: '#/components/schemas/ErrorModel'
        - type: object
          properties:
            rootCause:
              type: string
              example: Connexion to db failed
          required:
            - rootCause
      description: "Une message d'erreur avec la stack d'erreur"
      x-examples:
        example-1: {}
    ErrorModel:
      type: object
      description: "Un message d'erreur"
      x-examples:
        example-404:
          code: 404
          message: "Cette offre d'emploi n'existe pas."
      properties:
        message:
          type: string
          example: "Cette offre d'emploi n'existe pas"
          minLength: 10
        code:
          type: integer
          format: int32
          minimum: 200
          maximum: 500
      required:
        - message
        - code
    Identifier:
      title: Identifier
      type: object
      description: "L'identifiant unique des objets persistés dans le JobBoard"
      properties:
        identifier:
          type: string
          format: uuid
          example: a122edec-5580-4a93-aff7-fc18b41e4c57
...
    FullOrganization:
      title: FullOrganization
      description: Une entreprise complete avec identifiant unique
      allOf:
        - $ref: '#/components/schemas/Identifier'
        - $ref: '#/components/schemas/Organization'
      x-examples:
        example-1:
          identifier: a122edec-5580-4a93-aff7-fc18b41e4c57
          name: Incaya
          description: "INCAYA, société coopérative de services du numérique. Créons ensemble des outils qui ont du sens ! Collectivités, associations, entreprises : nous concevons avec vous des outils numériques qui transforment l'expérience de vos utilisateurs et usagers."
          image: 'https://www.incaya.fr/static/logo-incaya.svg'
          email: contact@organization.org
          url: 'https://www.incaya.fr/'
          address:
            addressCountry: France
            addressLocality: Caen
            postalCode: '14000'
            streetAddress: '5, place de la Répulique'
          contactPoints:
            - email: job@organization.org
              telephone: '0606060606'
              name: 'John Do, CTO'
              contactType: "Offres d'emploi"
    FullContactPoint:
      title: FullContactPoint
      allOf:
        - $ref: '#/components/schemas/Identifier'
        - $ref: '#/components/schemas/ContactPoint'
      x-examples:
        example-1:
          identifier: abeb9875-2191-4097-8249-5b8defc42a3e
          email: job@organization.org
          telephone: '0606060606'
          name: 'John Do, CTO'
          contactType: "Offres d'emploi"
    OrganizationForCreation:
      title: OrganizationForCreation
      type: object
      description: "Données nécessaires à la création d'une entreprise"
      properties:
        contactPoints:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/ContactPoint'
        address:
          anyOf:
            - type: string
            - $ref: '#/components/schemas/PostalAddress'
        image:
          type: string
          format: uri
          example: 'https://www.incaya.fr/static/logo-incaya.svg'
          minLength: 10
          pattern: "https?:\\/\\/(www\\.)?[-a-zA-Z0-9@:%._\\+~#=]{1,256}\\.[a-zA-Z0-9()]{1,6}\\b([-a-zA-Z0-9()@:%_\\+.~#?&//=]*)"
        email:
          type: string
          format: email
          example: email@organization.org
        url:
          type: string
          format: uri
          example: 'https://www.incaya.fr/'
          minLength: 10
          pattern: "https?:\\/\\/(www\\.)?[-a-zA-Z0-9@:%._\\+~#=]{1,256}\\.[a-zA-Z0-9()]{1,6}\\b([-a-zA-Z0-9()@:%_\\+.~#?&//=]*)"
        description:
          type: string
          example: Coopérative de services du numérique
        name:
          type: string
          example: Incaya
          minLength: 2
          maxLength: 150
      required:
        - contactPoints
        - url
        - name
  parameters:
    Filter:
      name: filters
      in: query
      description: "Filtres appliquables à la requête. C'est un objet json stringifié, avec les filtres en clé/valeur séparé par des virgules."
      required: false
      explode: true
      schema:
        type: string
    Sort:
      name: sort
      in: query
      description: "Le tri applicable à la liste. C'est un tableau stringifié de la forme [sortProp, sortDirection]"
      required: false
      explode: false
      schema:
        type: array
        maxItems: 2
        minItems: 2
        items:
          type: string
    PaginationCurrentPage:
      name: currentPage
      in: query
      description: "Le paramètre de pagination pour définir la page courrante."
      required: false
      schema:
        type: integer
        example: 1
    PaginationPerPage:
      name: perPage
      in: query
      description: "Le paramètre de pagination pour définir le nombre d'éléments par page."
      required: false
      schema:
        type: integer
        example: 10
    UUID:
      name: identifier
      in: path
      description: "Identifiant unique d'un objet de forme uuid"
      required: true
      schema:
        type: string
        format: uuid
      example: 9a6c8995-df54-446c-a5b8-71532c304751